+++ test/Problems/P70Spec.hs
@@ -6,8 +6,9 @@ Maintainer: dev@chungyc.org
 module Problems.P70Spec (spec) where
 
 import           Problems.MultiwayTrees
-import qualified Problems.P70           as Problem
-import qualified Solutions.P70          as Solution
+import           Problems.MultiwayTrees.QuickCheck
+import qualified Problems.P70                      as Problem
+import qualified Solutions.P70                     as Solution
 import           Test.Hspec
 import           Test.Hspec.QuickCheck
 import           Test.QuickCheck
@@ -32,13 +33,12 @@ properties
         multitreeToString t `shouldBe` [c] ++ concatMap multitreeToString ts ++ "^"
 
 examples :: Spec
-examples = do
-  describe "Examples" $ do
-    it "stringToMultitree \"afg^^c^bd^e^^^\" == multitree5" $ do
-      stringToMultitree "afg^^c^bd^e^^^" `shouldBe` multitree5
+examples = describe "Examples" $ do
+  it "stringToMultitree \"afg^^c^bd^e^^^\" == multitree5" $ do
+    stringToMultitree "afg^^c^bd^e^^^" `shouldBe` multitree5
 
-    it "multitreeToString multitree5" $ do
-      multitreeToString multitree5 `shouldBe` "afg^^c^bd^e^^^"
+  it "multitreeToString multitree5" $ do
+    multitreeToString multitree5 `shouldBe` "afg^^c^bd^e^^^"
 
   where stringToMultitree = Problem.stringToMultitree
         multitreeToString = Problem.multitreeToString
@@ -64,23 +64,5 @@ letters = choose ('a', 'z')
 newtype CharTree = CharTree (MultiwayTree Char) deriving (Show)
 
 instance Arbitrary CharTree where
-  arbitrary = CharTree <$> sized gen
-    where gen 0 = frequency
-                  [ (10, MultiwayTree <$> letters <*> return [])
-                  , (1, MultiwayTree <$> letters <*> subtreeList 0)
-                  ]
-          gen n = frequency
-                  [ (1, MultiwayTree <$> letters <*> return [])
-                  , (10, MultiwayTree <$> letters <*> subtreeList n)
-                  ]
-          subtreeList n = do
-            k <- chooseInt (0, n)
-            let s = case k of 0 -> 0; _ -> n `div` k
-            vectorOf k $ gen s
-
-  shrink (CharTree t) = map CharTree $ shrink' t
-    where shrink' (MultiwayTree _ []) = []
-          shrink' (MultiwayTree x xs) =
-            [ MultiwayTree x [] ] ++
-            xs ++
-            map (MultiwayTree x) (shrinkList shrink' xs)
+  arbitrary = CharTree <$> multiwayTreesOf letters
+  shrink (CharTree t) = map CharTree $ shrinkMultiwayTree shrinkNothing t