+++ test/Problems/P74Spec.hs
@@ -1,5 +1,5 @@
 {-|
-Copyright: Copyright (C) 2022 Yoo Chung
+Copyright: Copyright (C) 2023 Yoo Chung
 License: GPL-3.0-or-later
 Maintainer: dev@chungyc.org
 -}
@@ -18,22 +18,22 @@ import           Test.QuickCheck
 properties :: (Handle -> Handle -> IO ()) -> String -> Spec
 properties askGoldbach name = describe name $ do
   prop "prints Goldbach decomposition" $
-    \(Positive n') -> n' > 1 ==>
-                      let n = 2 * n' :: Int
-                          (a,b) = goldbach n
-                      in getOutput (\hOut -> withFixedInput (show n) hOut askGoldbach)
-                         `shouldReturn` (show n) ++ "=" ++ (show a) ++ "+" ++ (show b)
+    forAll (evenNumber `suchThat` (>2)) $ \n ->
+    getOutput (\hOut -> withFixedInput (show n) hOut askGoldbach)
+    `shouldReturn` printExpr (n, goldbach n)
+
+  where evenNumber = (2*) <$> arbitrarySizedNatural :: Gen Integer
+        printExpr (n, (a,b)) = (show n) ++ "=" ++ (show a) ++ "+" ++ (show b)
 
 examples :: Spec
-examples = do
-  describe "Examples" $ do
-    it "withFixedInput \"104\" stdout askGoldbach" $ do
-      getOutput (\hOut -> withFixedInput "104" hOut askGoldbach) `shouldReturn` "104=3+101"
+examples = describe "Examples" $ do
+  it "withFixedInput \"104\" stdout askGoldbach" $ do
+    getOutput (\hOut -> withFixedInput "104" hOut askGoldbach) `shouldReturn` "104=3+101"
 
-    it "withFixedInput \"104\" stdout askGoldbach'" $ do
-      getOutput (\hOut -> withFixedInput "104" hOut askGoldbach') `shouldReturn` "104=3+101"
+  it "withFixedInput \"104\" stdout askGoldbach'" $ do
+    getOutput (\hOut -> withFixedInput "104" hOut askGoldbach') `shouldReturn` "104=3+101"
 
-    where askGoldbach = Problem.askGoldbach
+  where askGoldbach = Problem.askGoldbach
 
 spec :: Spec
 spec = parallel $ do