+++ test/Problems/P76Spec.hs
@@ -1,5 +1,5 @@
 {-|
-Copyright: Copyright (C) 2022 Yoo Chung
+Copyright: Copyright (C) 2023 Yoo Chung
 License: GPL-3.0-or-later
 Maintainer: dev@chungyc.org
 -}
@@ -14,41 +14,42 @@ import           Test.Hspec.QuickCheck
 import           Test.QuickCheck
 import           Text.Read             (readMaybe)
 
-properties :: (String -> Either String (Int, (Int, Int))) -> String -> Spec
+properties :: (String -> Either String (Integer, (Integer, Integer))) -> String -> Spec
 properties eitherGoldbach name = describe name $ do
   prop "fails with non-number input" $
-    \s -> (isNothing (readMaybe s :: Maybe Int)) ==>
+    \s -> (isNothing (readMaybe s :: Maybe Integer)) ==>
     eitherGoldbach s `shouldBe` Left "not a number"
 
   prop "fails with number less than or equal to 2" $
-    \n -> (n :: Int) <= 2 ==>
+    \n -> (n :: Integer) <= 2 ==>
     eitherGoldbach (show n) `shouldBe` Left "not greater than 2"
 
   prop "fails with odd number" $
-    \n -> (n :: Int) > 2 && odd n ==>
+    forAll (oddNumbers `suchThat` (>2)) $ \n ->
     eitherGoldbach (show n) `shouldBe` Left "not an even number"
 
   prop "is same as goldbach with even number" $
-    \(Small n') -> n' > 2 ==>
-                   let n = 2 * n'
-                   in eitherGoldbach (show n) `shouldBe` Right (n, goldbach n)
+    forAll (evenNumbers `suchThat` (>2)) $ \n ->
+    eitherGoldbach (show n) `shouldBe` Right (n, goldbach n)
+
+  where oddNumbers = (\n -> n*2+1) <$> arbitrarySizedNatural :: Gen Integer
+        evenNumbers = (2*) <$> arbitrarySizedNatural
 
 examples :: Spec
-examples = do
-  describe "Examples" $ do
-    it "eitherGoldbach \"104\"" $ do
-      eitherGoldbach "104" `shouldBe` Right (104, (3,101))
+examples = describe "Examples" $ do
+  it "eitherGoldbach \"104\"" $ do
+    eitherGoldbach "104" `shouldBe` Right (104, (3,101))
 
-    it "eitherGoldbach \"this is not a number\"" $ do
-      eitherGoldbach "this is not a number" `shouldBe` Left "not a number"
+  it "eitherGoldbach \"this is not a number\"" $ do
+    eitherGoldbach "this is not a number" `shouldBe` Left "not a number"
 
-    it "eitherGoldbach \"1\"" $ do
-      eitherGoldbach "1" `shouldBe` Left "not greater than 2"
+  it "eitherGoldbach \"1\"" $ do
+    eitherGoldbach "1" `shouldBe` Left "not greater than 2"
 
-    it "eitherGoldbach \"101\"" $ do
-      eitherGoldbach "101" `shouldBe` Left "not an even number"
+  it "eitherGoldbach \"101\"" $ do
+    eitherGoldbach "101" `shouldBe` Left "not an even number"
 
-    where eitherGoldbach = Problem.eitherGoldbach
+  where eitherGoldbach = Problem.eitherGoldbach
 
 spec :: Spec
 spec = parallel $ do