+++ test/Problems/P77Spec.hs
@@ -1,5 +1,5 @@
 {-|
-Copyright: Copyright (C) 2022 Yoo Chung
+Copyright: Copyright (C) 2023 Yoo Chung
 License: GPL-3.0-or-later
 Maintainer: dev@chungyc.org
 -}
@@ -13,21 +13,21 @@ import           Test.Hspec.QuickCheck
 import           Test.QuickCheck
 
 properties :: (Int -> [[Int]]) -> String -> Spec
-properties randomWalkPaths name =
-  describe name $ modifyMaxSuccess (const 10) $ modifyMaxSize (const 10) $ do
+properties randomWalkPaths name = describe name $
+  modifyMaxSuccess (const 10) $
+  modifyMaxSize (const 10) $ do
 
-  prop "starts paths with 0" $
-    \(NonNegative n) -> randomWalkPaths n `shouldSatisfy` all ((==) 0 . head)
+  prop "has single path with zero steps" $
+    randomWalkPaths 0 `shouldBe` [[0]]
 
-  prop "steps changes position by one of [-1,0,1]" $
-    \(NonNegative n) -> randomWalkPaths n `shouldSatisfy`
-                        all (\l -> all (\x -> elem x [-1,0,1]) $ map (uncurry (-)) $ zip (tail l) l)
+  prop "changes position by one of [-1, 0, 1]" $ \(Positive n) ->
+    randomWalkPaths n `shouldMatchList`
+    concatMap
+    (\path -> [ path ++ [last path + step] | step <- [-1,0,1] ])
+    (randomWalkPaths $ n-1)
 
   prop "has paths with expected length" $
     \(NonNegative n) -> randomWalkPaths n `shouldSatisfy` all ((==) (n+1) . length)
-
-  prop "includes any random walk path" $ withMaxSuccess 100 $
-    \(NonNegative n) -> forAll (pathWith n) $ \p -> randomWalkPaths n `shouldSatisfy` elem p
 
 examples :: Spec
 examples = do
@@ -47,16 +47,3 @@ spec = parallel $ do
   examples
   describe "From solutions" $ do
     properties Solution.randomWalkPaths "randomWalkPaths"
-
--- Given the list of steps and path, return the path with the next path.
--- The list of steps are in order, but the path is in the reverse order of positions.
-walk :: [Int] -> [Int] -> [Int]
-walk [] path                      = path
-walk _ []                         = []
-walk (step:steps) path@(latest:_) = walk steps (latest+step : path)
-
--- Generator for a random walk path with n steps.
-pathWith :: Int -> Gen [Int]
-pathWith n = do
-  steps <- vectorOf n $ elements[-1,0,1]
-  return $ reverse $ walk steps [0]