+++ ChangeLog.md
@@ -9,6 +9,9 @@ See https://github.com/chungyc/ninetynine.
 *   Rename `Problems.PostfixNotation` to `Problems.Monads` to be consistent
     with other supporting modules.
 
+*   Remove `Duplicate` operator from `Problems.P79`, and also update
+    the problem to require a more complete calculation history.
+
 ## v1.0.0
 
 *   99 Haskell problems ready to see the light of day.